+++ CHANGELOG.md
@@ -2,9 +2,14 @@
 
 ## [Unreleased]
 
+### Added
+
+- Switching the languages in the review
+
 ### Fixed
 
 - Selecting same language for source and target in group creation
+- Last vocabulary in review session flips which leads to flickering
 
 ## 1.0.0 - 2026-09-15